test(observability): dedupe the fresh-handle test helper
freshHandleOver (the fake SqlExec-over-shared-storage simulating a
post-hibernation fresh isolate) was copy-pasted verbatim into both
function-metrics.test.ts and query-metrics.test.ts. Hoist it into
__tests__/_helpers/fresh-handle.ts, next to the existing node-sqlite
harness, and import it from both suites.
